Really fix # char handling
Pull #70 didn't work as intended. Sometimes # chars in inline code segments are shown literal and sometimes they are interpreted, causing the escaping \ to be visible. This is probably a bug in AsciiDoctor. As a workaround we could just avoid it altogether by replacing # by &num; in all inline text as is done in /source/tutorial/mosquitto.adoc.

Note that &num; is a new HTML entity that won't be recognized by ancient web browsers such as IE9 (according to https://stackoverflow.com/questions/3025171/whats-the-html-character-entity-for-the-sign/27091747#27091747).